Move WidgetUtils to utils dir
